requirements.txt
langchain==0.1.0
chromadb==0.4.22
sentence-transformers==2.2.2
openai==1.12.0
python-dotenv==1.0.0 